Add support for auto-detection of completion system.
Adds a new 'auto option to multi-compile-completion-system that will
attempt to determine the completion system enabled in the user's
environment, defaulting to the builtin completion framework when no
other completion system is detected.  This should prevent the need for
most users to even have to customize this variable, and will
automatically detect if the user changes completion frameworks.